Working with self-discovery based on the BLE Scan Response device name data
