--- config.json
@@ -1,11 +1,0 @@
-{
-  "name": "Hello world",
-  "version": "1",
-  "slug": "hello_world",
-  "description": "My first real add-on!",
-  "arch": ["armhf", "armv7", "aarch64", "amd64", "i386"],
-  "startup": "before",
-  "boot": "auto",
-  "options": {},
-  "schema": {}
-}
